+++ services/workflows-service/src/workflow/types/unified-callback-names.ts
@@ -2,4 +2,5 @@ export type UnifiedCallbackNames =
   | 'ubo-unified-api'
   | 'kyc-unified-api'
   | 'kyb-unified-api'
-  | 'kyc-unified-api-decision';
+  | 'kyc-unified-api-decision'
+  | 'website-monitoring';