+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # kyb-app
 
+## 0.3.0-11210907c.0
+
+### Minor Changes
+
+- Bump
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.0-11210907c.0
+  - @ballerine/ui@0.5.0-11210907c.0
+  - @ballerine/blocks@0.2.1-11210907c.0
+  - @ballerine/workflow-browser-sdk@0.6.1-11210907c.0
+
 ## 0.2.0
 
 ### Minor Changes
@@ -102,6 +116,10 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @ballerine/common@0.7.42-11210907c.0
+  - @ballerine/blocks@0.1.27-11210907c.0
+  - @ballerine/ui@0.3.29-11210907c.0
+  - @ballerine/workflow-browser-sdk@0.5.42-11210907c.0
   - @ballerine/common@0.7.42
   - @ballerine/workflow-browser-sdk@0.5.42